requirements.txt
pandas
requests
python-dotenv==0.21.0
langchain_openai
http
dashscope